refactor: AI 코드 리뷰 반영 - 컴포넌트 분리 및 최적화
- FavoritesList.tsx 컴포넌트 분리 (408줄 → 142줄)
  - FavoritesHeader, FavoritesEmptyState, FavoritesRow로 분리
- ScrollView와 FlatList 중첩 문제 해결
- FavoriteCounselorCard에 React.memo 적용
- useCallback으로 이벤트 핸들러 최적화
